DB check
There was no entry in the table t_quote due to an error in the content.sql error. As I have deleted some authors I forgot to adjust the fk in t_quote. The issue is now corrected.
I also adjusted some views - index and author - for a better presentation